% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbmodule-package.r
\docType{package}
\name{fbmodule}
\alias{fbmodule}
\alias{fbmodule-package}
\title{fbmodule.}
\description{
fbmodule.
}
\examples{
library(fbmodule)

# Create a dummy sample table

tbl <- get_module_table("potato")

}

