% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_module.R
\name{get_module_table}
\alias{get_module_table}
\title{Gets a site table.}
\usage{
get_module_table(crop = NULL)
}
\arguments{
\item{crop}{character}
}
\value{
dataframe
}
\description{
If not yet present creates a dummy one.
}
\author{
Reinhard Simon
}

