% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_module.R
\name{import_module_table}
\alias{import_module_table}
\title{Imports a module table locally.}
\usage{
import_module_table(file, crop)
}
\arguments{
\item{file}{a file path as string}

\item{crop}{character}
}
\value{
dataframe
}
\description{
Posts a data.frame containing location data and also returns the table or NULL.
}
\author{
Reinhard Simon
}

