% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_module.R
\name{list_modules}
\alias{list_modules}
\title{List modules for a crop}
\usage{
list_modules(crop)
}
\arguments{
\item{crop}{character}
}
\value{
vector of module names
}
\description{
List modules for a crop
}
\author{
Reinhard Simon
}

