% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_module.R
\name{new_module_table}
\alias{new_module_table}
\title{Creates an empty module table}
\usage{
new_module_table(crop = NULL)
}
\arguments{
\item{crop}{character}
}
\value{
dataframe
}
\description{
With dummy data
}
\author{
Reinhard Simon
}

