% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_module.R
\name{post_module_table}
\alias{post_module_table}
\title{Posts a module table locally.}
\usage{
post_module_table(table_module, crop)
}
\arguments{
\item{table_module}{data.frame}

\item{crop}{character}
}
\description{
Posts a data.frame containing location data.
}
\author{
Reinhard Simon
}

