% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_module.R
\name{server_module}
\alias{server_module}
\title{server_module}
\usage{
server_module(input, output, session, dom = "hot_modules", values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{dom}{target dom element name}

\item{values}{reactive values}
}
\description{
Constructs table
}
\author{
Reinhard Simon
}

