% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_module.R
\name{ui_module}
\alias{ui_module}
\title{shiny UI element}
\usage{
ui_module(type = "tab", title = "Module configuration",
  name = "resource_modules", output = "hot_modules")
}
\arguments{
\item{type}{of ui Element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}

\item{output}{name of output element}
}
\description{
returns a re-usable user interface element
}
\author{
Reinhard Simon
}

