% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ui_module_params.R
\name{ui_module_params}
\alias{ui_module_params}
\title{UI material list paramters}
\usage{
ui_module_params(name = "resource_modules")
}
\arguments{
\item{name}{character}
}
\description{
An interface to parameters
}
\author{
Reinhard Simon
}

